fishings: make data more consistent
The previous version had an issue with consistency. Since the Northern
and Southern hemisphere data was stored separately, there were times
when some of the info fields would differ by whitespace or
capitalization. Fixed that as well as changed the way months and times
are being stored. Front end was updated to reflect these changes.